Extending CPU dispatch to the tests; fixing a typo
